+++ clusterweaver/web/forms/__init__.py
@@ -1,5 +1,5 @@
 from clusterweaver.web.forms.nodes import NodeForm
-from clusterweaver.web.forms.projects import ProjectForm
+from clusterweaver.web.forms.projects import ProjectForm, ProjectImportForm
 from clusterweaver.web.forms.ssh import ConnectivityRunForm, HostsUpdateRunForm, NetworkApplyForm, NetworkCheckRunForm, PrecheckRunForm, SSHDiscoveryForm, SSHKeyBootstrapForm
 
-__all__ = ["ConnectivityRunForm", "HostsUpdateRunForm", "NetworkApplyForm", "NetworkCheckRunForm", "NodeForm", "PrecheckRunForm", "ProjectForm", "SSHDiscoveryForm", "SSHKeyBootstrapForm"]
+__all__ = ["ConnectivityRunForm", "HostsUpdateRunForm", "NetworkApplyForm", "NetworkCheckRunForm", "NodeForm", "PrecheckRunForm", "ProjectForm", "ProjectImportForm", "SSHDiscoveryForm", "SSHKeyBootstrapForm"]